+++ Thunderstore/CHANGELOG.md
@@ -1,7 +1,15 @@
+# 1.0.7  
+- added `kplayers` command.  
+
+# 1.0.6  
+- potential fix for exploded easter eggs showing up for `kscan`.  
+- added [Items being held] to `kitems`.  
+
 # 1.0.5  
 - added `kroes` command.  
 - added quota notification.  
 - added config.  
+- changed `kenemy` command.  
 
 # 1.0.4  
 - added `kenemy` command.